fix: ranged hit talent bonus for classic SOM
In classic the hit talent bonus is part of the return value of `GetHitModifier`. Since the function `GetHitTalentBonus` is only used for the TBC context I would suggest this easy fix, tough it is a bit confusing in a classic context as `GetHitTalentBonus` will return 0, eventough the 'correct' value would be `3`.

Also note that in classic the talent column would be `11` and not `12`, if you would rather go that route.

This change should also happen in the `Melee.lua` file.